requirements.txt
django < 4.0.0
django-mptt
django-mptt-admin
django-dotenv
dj_database_url
psycopg2-binary
django-debug-toolbar
# Uncomment the following line to work with django-query-profiler
# django-query-profiler
coverage